Create mapsRegulome Object

This function loads chromatin maps, obtains the classes present in coordinates and adds the needed parameters for plotting.

create_mapsRegulome(coordinates, maps_dataset, genome = "hg19",
  path = "~/data/IRB/")

Arguments

coordinates
Either a GRanges object, data.frame or character string (i.e. chr11:17227951-17589050) indicating the coordinates for the region the user wants to plot.

maps_dataset
Name of the chromatin maps to plot. The value should be one of the defined below (column Dataset). Defaults to "", which will produce an empty plot.

Name	Dataset	Reference

Adult Islets - Chromatin Classes	chromatinClassesReduced	Miguel-Escalada et al. (2019)

Adult Islets - Chromatin Classes	openChromatinClasses	Pasquali et al. (2014)

Pancreatic Progenitors	progenitors	Cebola et al. (2015)

Adult Islets - Chromatin States	chromatinStates	Parker et al. (2013)

genome
Character string indicating the genome for the coordinates. Default: hg19.

path
Path containing the genomes folder (for example "hg19"). Default: "~/data/IRB/"

Value

An object of class mapsRegulome, which is basically a list with the following attributes:

name: Name of selected chromatin map dataset.

col: Named vector with the color of each type of chromatin class.

genome: Genome build.

coordinates: Selected coordinates to plot as a GRanges object.

value: GRanges object containing coordinates and type of chromatin class.

moreArgs: Additional arguments, which are empty.

Examples

if (FALSE) {
maps <- create_mapsRegulome(coordinates="chr20:22461643-22666093",
                            maps_dataset="openChromatinClasses",
                            path="~/data/IRB_database/")
}