Create snpsRegulome Object

This function loads the GWAS information for SNPs from the specified dataset, selects those present in coordinates and adds all the elements needed for plotting.

create_snpsRegulome(coordinates, snps_dataset = "",
  snps_col = "dark red", genome = "hg19", top = 3,
  snps_scaling = NULL, path = "~/data/IRB/")

Arguments

coordinates
Either a GRanges object, data.frame or character string (i.e. chr11:17227951-17589050) indicating the coordinates for the region the user wants to plot.

snps_dataset
GWAS SNPs dataset name to use for the analysis. The value should be one of the defined below (column Dataset). Defaults to "", which will produce an empty plot.

Name	Dataset	Reference

70KforT2D	70KforT2D	S. Bonàs-Guarch et al. (2018)

Diagram	diagram	A. P. Morris et al. (2012)

Magic	magic	R. A. Scott et al. (2012)

snps_col
Color for the GWAS SNPs. Default: "dark red"

genome
Character string indicating the genome for the coordinates. Default: hg19.

top
Number indicating how many of the most significant SNPs in the region will be shown as labels.

snps_scaling
Number indicating the value to which scale the maximum -log10 PVAL for SNPs. Defaults to NULL, meaning that no scaling will be performed.

path
Path containing the genomes folder (for example "hg19"). Default: "~/data/IRB/"

Value

An object of class snpsRegulome, which is basically a list with the following attributes:

name: Name of the selected SNP dataset.

col: Color for plotting the SNPs.

genome: Genome build.

coordinates: Selected coordinates to plot as a GRanges object.

value: GRanges object containing all SNP values in coordinates.

moreArgs: Additional arguments, which are:

top: contains the number of SNPs with lower P-values whose name will be plot

scaleTo: parameter that includes a scaling parameter in case virtual4C is also plot along SNPs (it is the maximum virtual4C value and if no virtual4C is plot, is the maximum SNP -log10 p-value).

Examples

if (FALSE) {
snps <- create_snpsRegulome(coordinates="chr20:22461643-22666093",
                            snps_dataset="diagram",
                            path="~/data/IRB_database/")
}