plotRegulome package

Plotting Islet Regulome data

By providing the coordinates and IRB datasets of interest, it will produce the characteristic Islet Regulome Broser plot.

plotRegulome(coordinates, snps_dataset = "", snps_col = "dark red",
  contacts_dataset = "", contacts_col = c(`0` = "grey", `3` = "blue",
  `5` = "dark orange"), maps_dataset = "", clusters_dataset = "",
  cluster_col = "dark green", tfs_dataset = "",
  tfs_col = "dark blue", genes_col = c(gene = "dark grey", spec =
  "darkorchid3", lnc = "black"), showLongestTranscript = TRUE,
  randomIRB = FALSE, genome = "hg19", path = "~/data/IRB/")

Arguments

coordinates
Either a GRanges object, data.frame or character string (i.e. chr11:17227951-17589050) indicating the coordinates for the region the user wants to plot.

snps_dataset
GWAS SNPs dataset name to use for the analysis. The value should be one of the defined below (column Dataset). Defaults to "", which will produce an empty plot.

Name	Dataset	Reference

70KforT2D	70KforT2D	S. Bonàs-Guarch et al. (2018)

Diagram	diagram	A. P. Morris et al. (2012)

Magic	magic	R. A. Scott et al. (2012)

snps_col
Color for the GWAS SNPs. Default: "dark red"

contacts_dataset
BaitID or bait gene name for the virtual 4C data (Miguel-Escalada et al. (2019)) to plot.

contacts_col
Named vector ("0", "3" and "5"), with the colors for each group of CHiCAGO contacts.

maps_dataset
Name of the chromatin maps to plot. The value should be one of the defined below (column Dataset). Defaults to "", which will produce an empty plot.

Name	Dataset	Reference

Adult Islets - Chromatin Classes	chromatinClassesReduced	Miguel-Escalada et al. (2019)

Adult Islets - Chromatin Classes	openChromatinClasses	Pasquali et al. (2014)

Pancreatic Progenitors	progenitors	Cebola et al. (2015)

Adult Islets - Chromatin States	chromatinStates	Parker et al. (2013)

clusters_dataset
Name of the enhncer cluster data to plot. The value should be one of the defined below (column Dataset). Defaults to "", which will produce an empty plot.

Name	Dataset	Reference

Enhancer Hubs	enhancerHubs	Miguel-Escalada et al. (2019)

Super Enhancers	superEnhancers	Miguel-Escalada et al. (2019)

Enhancer Clusters	enhancerClusters	Pasquali et al. (2014)

Stretch Enhancers	stretchEnhancers	Parker et al. (2013)

COREs	cores	K. J. Gaulton et al. (2010)

cluster_col
Color for the clusters. Default: "dark green"

tfs_dataset
Name of the TF dataset to plot. The value should be one of the defined below (column Dataset). Defaults to "", which will produce an empty plot.

Name	Dataset	Reference

Adult Islets – Structural	structure	Miguel-Escalada et al. (2019)

Pancreatic Progenitors	progenitors	Cebola et al. (2015)

Adult Islets – Tissue-specific	adult	Pasquali et al. (2014)

tfs_col
Color for the TFs circle border.

genes_col
Named character vector ("gene", "lnc" and "spec") with the colors for each type of feture plotted in the gene annotation track.

showLongestTranscript
When plotting gene data, set to TRUE (default) if you want to reduce the number of transcripts by only plotting the longest transcript per gene. If set to FALSE, will plot all available transcripts.

randomIRB
Generate random combinations of IRB datasets.

genome
Character string indicating the genome for the coordinates. Default: hg19.

path
Path containing the genomes folder (for example "hg19"). Default: "~/data/IRB/"

Value

A ggplot2 object that can be plotted or saved with ggsave. Recomended device size in inches is: width=11, height=6

Details

See the README on GitHub

Examples

if (FALSE) {
    coordinates <- "chr5:131817301-131826490"
    snps_dataset <- "70KforT2D"
    contacts_dataset <- 570519 #183446
    maps_dataset <- "progenitors"
    clusters_dataset <- ""
    tfs_dataset <- ""
    path <- "~/data/IRB/"

    plotRegulome(coordinates=coordinates,
                 snps_dataset=snps_dataset,
                 contacts_dataset=contacts_dataset,
                 maps_dataset=maps_dataset,
                 clusters_dataset=clusters_dataset,
                 tfs_dataset=tfs_dataset,
                 path=path)
}